Add inline caching for LOAD_GLOBAL and LOAD_ATTR
Summary:
This diff adds a poor man's version of inline caching [1] for LOAD_ATTR
and LOAD_GLOBAL. When a code object becomes "hot enough," we initialize
a cache that lives alongside its bytecode. Once initialized, we check the
cache before executing certain instructions. On a cache miss, we fall
back to the slow path and update the cache with the result.

The LOAD_GLOBAL opcode loads the value bound to a global name onto
the value stack. To achieve that, it performs up to two dictionary
lookups. It first looks in the `globals` dictionary attached to the
current stack frame. If that fails, it then looks in the `builtins`
dictionary of the frame. The `globals` and `builtins` dictionaries
rarely change, so caching the result should be an effective
optimization.

The LOAD_ATTR opcode looks up a named attribute on a receiver object
at the top of the operand stack. The process proceeds in the following
order:

1. Look for the attribute in each class in the receiver's MRO.
2. If the result from (1) is a data descriptor, invoke it and return
   the result.
3. Look up the attribute in the receiver's instance dictionary.
4. If (3) was successful, return the result.
5. If the result from (1) is a non-data descriptor, invoke it and
   return the result.
6. If the result from (1) is not `NULL`, return it.
7. Raise an error.

Thankfully, Python versions classes and invalidates the versions
down the subclass hierarchy whenever a parent class changes. This allows
us to cache the result of the MRO lookup.

The MRO walk is particularly expensive for instance attribute access.
It's almost always a miss and adds N additional dictionary lookups for
each attribute access.

Cache profiling infrastructure is provided via the `INLINE_CACHE_PROFILE`
macro. When defined, we record the following basic statistics about the
state of inline caches:

- Total number of caches allocated
- Total size of all caches

On a per-opcode basis, we record:

- Number of hits
- Number of misses
- Number of uncacheable call sites
- Number of entries

A 1 hour sample in production produces the following average stats:

- Total number of caches: 6279
- Total size: 2.68M
- LOAD_GLOBAL hit rate: 0.9263
- LOAD_ATTR hit rate: 0.9735

[1]: https://en.wikipedia.org/wiki/Inline_caching

Performance wise, this shows a 2% reduction of CPU instructions retired
on Instagram's workload.