Notify done_event after tstate has been unlinked.
Threading shutdown waits for all non-daemon threads to exit. In practice,
this means that we expect that after `threading._shutdown` completes,
there are no non-daemon threads linked into the interpreter.

Shutdown waits for the `done_event` for each non-daemon thread to be
set before returning. Previously, the `done_event` was set right at
the end of `PyThreadState_Clear` but before the `PyThreadState` is
unlinked from the interpreter in `PyThreadState_Delete{Current}`.

This is not an issue when the GIL is enabled. At the end of a thread's
life (in thread_run) a call to `PyThreadState_Clear` is followed
immediately by a call to `PyThreadState_DeleteCurrent`. The GIL ensures
that the notification in `PyThreadState_Clear` and the removal of the
`PyThreadState` in `PyThreadState_DeleteCurrent` happen atomically
from the perspective of `threading._shutdown`.

Unfortunately this is racy when the GIL is disabled. `threading._shutdown`
could return before the PyThreadState has been unlinked by `PyThreadState_DeleteCurrent`.