Don't promote borrowed references in STORE_FAST
This reduces the number of changes to the interpreter loop and leaves
the various STORE_FAST instructions unmodified. It means that we cannot
optimize LOAD_FAST instructions that are consumed by STORE_FAST instructions,
but I think this is a net positive. It speeds up STORE_FAST and doesn't appreciably
reduce the number of optimized instructions in the benchmark suite.